botMain/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
protobuf==3.19.4
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
